Move telemetry init stuff to API crate
