+++ core-api/Cargo.toml
@@ -13,7 +13,6 @@ categories = ["development-tools"]
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-anyhow = "1.0"
 async-trait = "0.1"
 derive_builder = "0.11"
 opentelemetry = "0.18"
@@ -22,6 +21,7 @@ serde_json = "1.0"
 thiserror = "1.0"
 tonic = "0.8"
 tracing-core = "0.1"
+url = "2.3"
 
 [dependencies.temporal-sdk-core-protos]
 path = "../sdk-core-protos"
